Update java files

+++ src/main/java/com/my/food_ordering/model/Order.java
@@ -42,7 +42,9 @@ public class Order {
 
     private int totalItem;
 
-    private int totalPrice;
+    private Long totalPrice;
+
+
 
     public Long getId() {
         return id;
@@ -116,11 +118,11 @@ public class Order {
         this.totalItem = totalItem;
     }
 
-    public int getTotalPrice() {
+    public Long getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(Long totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
